backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
aiohttp
beautifulsoup4
lxml
